+++ CHANGELOG.md
@@ -185,6 +185,7 @@ All notable changes to this project will be documented in this file.
 - Prevent endless loop in LogarithmicAxis (#957)
 - Fixed WPF series data not refreshed when not visible (included WPF LiveDemo)
 - Fixed bug in selection of plot to display in OxyPlot.GtkSharp ExampleBrowser (#979)
+- Fixed non-interpolation of HeatMapSeries in OxyPlot.GtkSharp (#980)
 
 ## [2014.1.546] - 2014-10-22
 ### Added